+++ WebApi/InWords.WebApi/InWords.WebApi/Modules/WordsSets/InModuleInitializer.cs
@@ -15,6 +15,7 @@ namespace InWords.WebApi.Modules.WordsSets
             builder.RegisterType<GetLevelWords>().AsImplementedInterfaces().InstancePerDependency();
             builder.RegisterType<GetTrainingLevelsHistory>().AsImplementedInterfaces().InstancePerDependency();
             builder.RegisterType<EstimateTraining>().AsImplementedInterfaces().InstancePerDependency();
+            builder.RegisterType<GetFullSets>().AsImplementedInterfaces().InstancePerDependency();
         }
     }
 }